2021-05-27

+++ modules/shop/item/set/controller/delete.php
@@ -9,7 +9,7 @@ defined('HOSTCMS') || exit('HostCMS: access denied.');
  * @subpackage Shop
  * @version 6.x
  * @author Hostmake LLC
- * @copyright © 2005-2019 ООО "Хостмэйк" (Hostmake LLC), http://www.hostcms.ru
+ * @copyright © 2005-2021 ООО "Хостмэйк" (Hostmake LLC), http://www.hostcms.ru
  */
 class Shop_Item_Set_Controller_Delete extends Admin_Form_Action_Controller
 {
@@ -21,16 +21,19 @@ class Shop_Item_Set_Controller_Delete extends Admin_Form_Action_Controller
 	{
 		$set_item_id = intval(Core_Array::getGet('set_item_id'));
 
-		$oShop_Item_Set = Core_Entity::factory('Shop_Item_Set')->getById($set_item_id);
-
-		if (!is_null($oShop_Item_Set))
+		if ($set_item_id)
 		{
-			$oShop_Item_Set->delete();
-		}
+			$oShop_Item_Set = Core_Entity::factory('Shop_Item_Set')->getById($set_item_id);
 
-		$this->_Admin_Form_Controller->addMessage(
-			"<script>$('.set-item-table tr#{$set_item_id}').remove();</script>"
-		);
+			if (!is_null($oShop_Item_Set))
+			{
+				$oShop_Item_Set->delete();
+			}
+
+			$this->_Admin_Form_Controller->addMessage(
+				"<script>$('.set-item-table tr#{$set_item_id}').remove();</script>"
+			);
+		}
 
 		return TRUE;
 	}
