+++ modules/shop/warehouse/incoming/item/controller/delete.php
@@ -9,7 +9,7 @@ defined('HOSTCMS') || exit('HostCMS: access denied.');
  * @subpackage Shop
  * @version 6.x
  * @author Hostmake LLC
- * @copyright © 2005-2020 ООО "Хостмэйк" (Hostmake LLC), http://www.hostcms.ru
+ * @copyright © 2005-2021 ООО "Хостмэйк" (Hostmake LLC), http://www.hostcms.ru
  */
 class Shop_Warehouse_Incoming_Item_Controller_Delete extends Admin_Form_Action_Controller
 {
@@ -21,36 +21,39 @@ class Shop_Warehouse_Incoming_Item_Controller_Delete extends Admin_Form_Action_C
 	{
 		$shop_warehouse_incoming_item_id = intval(Core_Array::getGet('shop_warehouse_incoming_item_id'));
 
-		$oShop_Warehouse_Incoming_Item = Core_Entity::factory('Shop_Warehouse_Incoming_Item')->getById($shop_warehouse_incoming_item_id);
-
-		if (!is_null($oShop_Warehouse_Incoming_Item))
+		if ($shop_warehouse_incoming_item_id)
 		{
-			$oShop_Warehouse_Incoming = $oShop_Warehouse_Incoming_Item->Shop_Warehouse_Incoming;
-			$oShop_Warehouse = $oShop_Warehouse_Incoming->Shop_Warehouse;
+			$oShop_Warehouse_Incoming_Item = Core_Entity::factory('Shop_Warehouse_Incoming_Item')->getById($shop_warehouse_incoming_item_id);
 
-			$shop_item_id = $oShop_Warehouse_Incoming_Item->shop_item_id;
-
-			$oShop_Warehouse_Incoming_Item->delete();
-
-			// Удаляем проводки в документе
-			$aShop_Warehouse_Entries = $oShop_Warehouse->Shop_Warehouse_Entries->getByDocumentAndShopItem($oShop_Warehouse_Incoming->id, $oShop_Warehouse_Incoming::TYPE, $shop_item_id);
-			foreach ($aShop_Warehouse_Entries as $oShop_Warehouse_Entry)
+			if (!is_null($oShop_Warehouse_Incoming_Item))
 			{
-				$oShop_Warehouse_Entry->delete();
+				$oShop_Warehouse_Incoming = $oShop_Warehouse_Incoming_Item->Shop_Warehouse_Incoming;
+				$oShop_Warehouse = $oShop_Warehouse_Incoming->Shop_Warehouse;
+
+				$shop_item_id = $oShop_Warehouse_Incoming_Item->shop_item_id;
+
+				$oShop_Warehouse_Incoming_Item->delete();
+
+				// Удаляем проводки в документе
+				$aShop_Warehouse_Entries = $oShop_Warehouse->Shop_Warehouse_Entries->getByDocumentAndShopItem($oShop_Warehouse_Incoming->id, $oShop_Warehouse_Incoming::TYPE, $shop_item_id);
+				foreach ($aShop_Warehouse_Entries as $oShop_Warehouse_Entry)
+				{
+					$oShop_Warehouse_Entry->delete();
+				}
+
+				// Удаляем все накопительные значения с датой больше, чем дата документа
+				Shop_Warehouse_Entry_Accumulate_Controller::deleteEntries($shop_item_id, $oShop_Warehouse->id, $oShop_Warehouse_Incoming->datetime);
+
+				$rest = $oShop_Warehouse->getRest($shop_item_id);
+
+				// Recount
+				$oShop_Warehouse->setRest($shop_item_id, is_null($rest) ? 0 : $rest);
 			}
 
-			// Удаляем все накопительные значения с датой больше, чем дата документа
-			Shop_Warehouse_Entry_Accumulate_Controller::deleteEntries($shop_item_id, $oShop_Warehouse->id, $oShop_Warehouse_Incoming->datetime);
-
-			$rest = $oShop_Warehouse->getRest($shop_item_id);
-
-			// Recount
-			$oShop_Warehouse->setRest($shop_item_id, is_null($rest) ? 0 : $rest);
+			// $this->_Admin_Form_Controller->addMessage(
+				// "<script>$('.shop-item-table tr#{$oShop_Warehouse_Incoming_Item->id}').remove();</script>"
+			// );
 		}
-
-		// $this->_Admin_Form_Controller->addMessage(
-			// "<script>$('.shop-item-table tr#{$oShop_Warehouse_Incoming_Item->id}').remove();</script>"
-		// );
 
 		return TRUE;
 	}
